Add a non-random robot info getter
Allow an energy bonus as part of the robot name info.  Barcodes that are
more difficult to create should yield robots with higher starting
energy.

Robocode now accepts barcodes of the form \d\d\d-\d\d\d-\d\d-IDX to
specify an adjective and noun from the respective dictionaries.  The 3rd
number is to specify the number of the robot AI.  Robots created this
way get 40 extra hitpoints.